+++ app/routes/components/Button/changelog.codegen.ts
@@ -21,4 +21,5 @@ export const changelog = ''
   + '  default icon size for SVG children in left or right section.\n'
   + '- 2025-11-15: Update icon size to make it smaller, remove the dependency on\n'
   + '  deprecated ThemeProvider.\n'
+  + '- 2025-11-23: Update the CSS variables names with camelCase.\n'
   ;