+++ app/routes/components/IconButton/changelog.codegen.ts
@@ -7,4 +7,5 @@ export const changelog = ''
   + '- 2025-09-26: Use theme tokens for styling.\n'
   + '- 2025-10-11: Code formatting improvements.\n'
   + '- 2025-10-30: Add the default icon size for SVG children.\n'
+  + '- 2025-11-23: Update the CSS variables names with camelCase.\n'
   ;