+++ src/components/ButtonGroup/index.ts
@@ -1,4 +1,4 @@
-// ButtonGroup from copy-ui @ 2025-10-10.
+// ButtonGroup from copy-ui @ 2025-11-23.
 
-export type { ButtonGroupProps } from "./ButtonGroup";
-export { ButtonGroup } from "./ButtonGroup";
+export type { ButtonGroupProps } from "./button-group";
+export { ButtonGroup } from "./button-group";