+++ src/components/ButtonGroup/CHANGELOG.md
@@ -1,3 +1,4 @@
 - 2025-09-09: Initial version.
 - 2025-09-30: Support `className` prop for custom styling.
 - 2025-10-10: Code formatting improvement.
+- 2025-11-23: Update the CSS variables names with camelCase.