+++ src/components/IconButton/CHANGELOG.md
@@ -3,3 +3,4 @@
 - 2025-09-26: Use theme tokens for styling.
 - 2025-10-11: Code formatting improvements.
 - 2025-10-30: Add the default icon size for SVG children.
+- 2025-11-23: Update the CSS variables names with camelCase.